Remove dependency to tkinter in progress bar builder
